requirements.txt
langchain
openai
faiss-cpu
streamlit
pandas
